Grader App
This app displays an interface in which the user can add either a grader or a graduate grader. Upon selection, the program asks for grader id, first name, last name, hourly pay rate, and the number of hours worked. If a graduate grader is selected, there is a text box to input the graduate grader's stipend as well. Upon inputting this information and clicking add grader, a message box appears which shows the grader's total pay. Upon closing this message box, the grader's information is displayed in the original interface.