Ticket App
On the main form of the app there is an option to add a new ticket, cancel, or to find out who the developer is. When the user selects add a ticket, a new form is brought up with the following information regarding the ticket: id, name, price, and process fee. If the developer button is clicked, the app displays who the developer is.